apps/api/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "asklio-api"
version = "0.1.0"
description = "askLio API"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.6",
    "python-dotenv>=1.0.1",
    "uvicorn>=0.34.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/asklio_api"]

[dependency-groups]
dev = [
    "mypy>=1.14.0",
    "pytest-asyncio>=0.25.0",
    "pytest>=8.3.4",
    "ruff>=0.8.4",
    "httpx>=0.28.1",
]

[tool.mypy]
strict = true

[[tool.mypy.overrides]]
module = ["tests/*"]
disable_error_code = ["no-untyped-def"]

[[tool.mypy.overrides]]
module = "dotenv"
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
